+++ dejafu-tests/Cases.hs
@@ -4,6 +4,7 @@ import Test.Framework (Test, testGroup)
 
 import qualified Cases.SingleThreaded as S
 import qualified Cases.MultiThreaded  as M
+import qualified Cases.Refinement     as R
 import qualified Cases.Litmus         as L
 import qualified Cases.Async          as A
 
@@ -12,6 +13,7 @@ testCases :: [Test]
 testCases = map (uncurry testGroup)
   [ ("Single Threaded", S.tests)
   , ("Multi Threaded",  M.tests)
+  , ("Refinement",      R.tests)
   , ("Litmus",          L.tests)
   , ("Async",           A.tests)
   ]